WIP  * Create order by param class to order runs  * Fix RunDto, by replacing property game for GameEmbedDto  * Add null as default parameter because in some results exception happenned(still trying to understand why)

+++ app/src/main/java/com/dluche/myspeedrunners/data/datasource/model/run/TimesDto.kt
@@ -9,17 +9,17 @@ data class TimesDto(
     @SerialName("ingame")
     val ingame: String?,
     @SerialName("ingame_t")
-    val ingameT: Int?,
+    val ingameT: Float?,
     @SerialName("primary")
     val primary: String?,
     @SerialName("primary_t")
-    val primaryT: Int?,
+    val primaryT: Float?,
     @SerialName("realtime")
     val realtime: String?,
     @SerialName("realtime_noloads")
     val realtimeNoloads: String?,
     @SerialName("realtime_noloads_t")
-    val realtimeNoloadsT: Int?,
+    val realtimeNoloadsT: Float?,
     @SerialName("realtime_t")
-    val realtimeT: Int?
+    val realtimeT: Float?
 )
